refactor SAST to remove blocks
Blocks are a detail of the source language that affect variable scope
and are otherwise unneccesary. Variables are resolved in semantic
checking, so variable scope is irrelevant in the SAST, and so having
blocks that can nest is unnecessary; just make if statements and loops
contain a flat list of statements, which we were effectively doing
anyways in order to make sure that breaks, continues, and returns were
the last statements.