add support for variable scoping and initializers
Now, we can declare variables anywhere, and optionally initialize them
as well. The scope handling is done immutably, which actually simplifies
things compared to Edward's slides; any changes made to the symbol table
are automatically undone when we return from check_stmt, so we don't
need a stack of variable bindings at all.