check for recursive function calls in GPU code
GLSL only lets functions call earlier functions like C, and doesn't
support forward declarations, which is how it effectively disallows
recursion. We allow functions to be defined in any order, but we still
have to disallow recursion somehow. We do this by doing a topological
sort, which also conveniently gives us an order for the functions that
makes GLSL happy. Do the same for structures, since GLSL has the same
problem there.