Extension of binary comparators to matrix types
Addition of general matrix matrix comparators. The philosophy is to return
a single boolean in all cases if the condition holds true for all
coefficients, coefficient wise. The exception is logical && and
logical || for boolean vector types, which returns a boolean vector
with the coefficient wise operation performed for each coef.